fix: XA decoder — match DuckStation formula exactly
- Use int16 cast before arithmetic shift (DuckStation Truncate16 pattern)
- Use s32 for prev samples (not s16) to match DuckStation accumulator
- Use std::clamp for sample clamping
- Remove redundant state save (prev pointer writes directly)

+++ src/audio/xa_decoder.h
@@ -62,9 +62,9 @@ class XaDecoder
     uint8_t filter_file_{0};
     uint8_t filter_channel_{0};
 
-    // ADPCM decode state (separate for left/right)
-    int16_t prev_left_[2]{0, 0};
-    int16_t prev_right_[2]{0, 0};
+    // ADPCM decode state — s32 like DuckStation (separate for left/right)
+    int32_t prev_left_[2]{0, 0};
+    int32_t prev_right_[2]{0, 0};
 
     // Last sector info
     int sample_rate_{37800};
